embedder: split into independent mac and winit embedders
Remove the embedder-backend selection shim and the embedder-core grab-bag.
The embedder layer is now three crates sharing only the webview crate API:

- embedder (root): thin dispatcher — CLI entry points and the compile-time
  windowed-backend selection (AppKit on macOS by default, winit elsewhere,
  winit forced via the winit_embedder feature).
- mac-embedder: self-contained AppKit app with its own user-event bus and
  platform helpers; no winit, Blitz, or GPU dependencies.
- winit-embedder: windowed app (Blitz chrome) plus the headless app used by
  WPT/WebDriver/CDP. The windowed app is gated behind a `windowed` feature;
  on macOS the crate builds headless-only by default, so the default build
  drops wgpu/Blitz/vello-cpu entirely.

Options are passed as parameters instead of a global static, embedders use
webview::NavigationCompleted directly, and the Vello-CPU screenshot path is
replaced with a plain white PNG. WebDriver/CDP remain available on every
configuration. AGENTS.md notes that testing builds the winit headless app.
