--- embedder/backend/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "embedder-backend"
-version = "0.1.0"
-edition = "2024"
-publish = false
-
-[features]
-default = []
-# On macOS the winit windowed embedder is not compiled unless this feature
-# is enabled (the AppKit backend is the default there). On other platforms
-# the winit backend is the only option and this feature is a no-op.
-winit_embedder = ["dep:winit-embedder"]
-
-[target.'cfg(target_os = "macos")'.dependencies]
-embedder-core = { path = "../core" }
-mac-embedder = { path = "../mac-embedder" }
-winit-embedder = { path = "../winit-embedder", optional = true }
-
-[target.'cfg(not(target_os = "macos"))'.dependencies]
-embedder-core = { path = "../core" }
-winit-embedder = { path = "../winit-embedder" }
