+++ embedder/mac-embedder/Cargo.toml
@@ -4,14 +4,19 @@ version = "0.1.0"
 edition = "2024"
 publish = false
 
+[features]
+default = []
+media = ["webview/media"]
+
 [dependencies]
+arboard = { version = "3.6.1", default-features = false }
 automation = { path = "../../automation" }
-embedder-core = { path = "../core" }
 ipc = { path = "../../ipc" }
 ipc-channel = { git = "https://github.com/gterzian/ipc-channel", rev = "169da83" }
 ipc_messages = { path = "../../ipc_messages" }
 keyboard-types = "0.7"
 log = "0.4"
+png = "0.17"
 serde_json = "1"
 uuid = { version = "1", features = ["v4"] }
 verification = { path = "../../verification" }
